Nightly commit
This commit contains some (most?) of the new authentication scheme necessary for
supporting client-side encryption. Presently, salts are retrieved from the
backend using an AJAX request. The intention from here is to use the
authentication salt with PBKDF2 and then populate and submit a hidden form with
the updated authentication key (effectively the password) and the username. This
commit also adds a bunch of Node files to support CryptoJS. There is some
question of whether CryptoJS can be used for client side work or just server
side, so these files may be removed in the future.