Nightly commit
This commit adds a function `f()` (which needs to be named appropriately) which
converts a Javascript ArrayBuffer into a string (by converting each byte into
a unit8 and then concatenating all of the together) so that it can be submitted
in a form as a password. This commit also contains a bunch of other code I was
playing around with which needs to be deleted in a subsequent commit, but right
now I'm tired.