Stop tracking minified JS and CSS
In the discussion surrounding
https://github.com/loboyd/setonotes/pull/6, loboyd and I determined that
tracking minified JavaScript and CSS is a bit redundant and isn't much
better than tracking a compiled binary. This commit implements that
determination by removing the minified files and adding them to
.gitignore. Actual deployments of `setonotes` will need to generate
these files and put them in the appropriate places, either manually or
by use of the build script.