Implement new auth (un-tuned)
This commit completes the first working version the new authentication scheme.
This includes generating new account's salts (but NOT encryption keys) on the
client. Encryption keys will be generated on the client when the other
encryption functionality is moved to the client. Still left to be done for auth
is determining appropriate salt lengths and tuning PBKDF2 to an appropriate
number of iterations.