Fix .gitignore typo and improve build script
There was a typo in .gitignore that didn't properly ignore
easymde.min.js. That's fixed now.

Since the minified files are ignored, the static/ directory isn't
currently being tracked. In the future, this will probably change as we
add more static files, but for now there's a possibility that the local
copy of the repo won't have a static/ tree. This can cause the build
script to fail because there's no place for the script to put the
minified CSS and JS files. The build script now accounts for this case.