Improve editor loading experience
This commit hides or removes the form elements that the editor uses to
communicate with the server. The editor experience is improved because
these elements no longer appear in the moment between initial page load
and when the editor actually appears on top of them. The navbar has been
removed for the same reason.