Refactoring the Editor to Be More Modular (#7)
* Fix .gitignore typo and improve build script

There was a typo in .gitignore that didn't properly ignore
easymde.min.js. That's fixed now.

Since the minified files are ignored, the static/ directory isn't
currently being tracked. In the future, this will probably change as we
add more static files, but for now there's a possibility that the local
copy of the repo won't have a static/ tree. This can cause the build
script to fail because there's no place for the script to put the
minified CSS and JS files. The build script now accounts for this case.

* Move custom editor code into its own file

* Improve editor loading experience

This commit hides or removes the form elements that the editor uses to
communicate with the server. The editor experience is improved because
these elements no longer appear in the moment between initial page load
and when the editor actually appears on top of them. The navbar has been
removed for the same reason.

* Rename setonotes.js to editor.js