Implement authentication compatible with client-side encryption (#8)
* Update /signin/ styling

* Give /signin/ a link home

* Nightly commit

This commit contains some (most?) of the new authentication scheme necessary for
supporting client-side encryption. Presently, salts are retrieved from the
backend using an AJAX request. The intention from here is to use the
authentication salt with PBKDF2 and then populate and submit a hidden form with
the updated authentication key (effectively the password) and the username. This
commit also adds a bunch of Node files to support CryptoJS. There is some
question of whether CryptoJS can be used for client side work or just server
side, so these files may be removed in the future.

* Give hidden inputs IDs

* Add PBKDF2 to auth

* Nightly commit

This commit adds a function `f()` (which needs to be named appropriately) which
converts a Javascript ArrayBuffer into a string (by converting each byte into
a unit8 and then concatenating all of the together) so that it can be submitted
in a form as a password. This commit also contains a bunch of other code I was
playing around with which needs to be deleted in a subsequent commit, but right
now I'm tired.

* Clean up new signin code

* Convert signup to use new authentication scheme

This commit does not complete this functionality. Stay tuned.

* Implement new auth (un-tuned)

This commit completes the first working version the new authentication scheme.
This includes generating new account's salts (but NOT encryption keys) on the
client. Encryption keys will be generated on the client when the other
encryption functionality is moved to the client. Still left to be done for auth
is determining appropriate salt lengths and tuning PBKDF2 to an appropriate
number of iterations.

* Capture ENTER key for signin/signup

Also, remove a `node_modules` directory containing `crypto-js` stuff that won't
be used.

* Clean up auth for develep merge

* Add udpated database schema

* Add doc-string to saltsHandler

* Remove unnecessary files