Make the first line the title of the note
Before this commit, there was no easy way to set or change the title of
a note using the new editor. Initially, I thought it might be a good
idea to add an input field somewhere so that the title can exist
separately from the note.  Later, I came to the conclusion that it would
be simpler to just have the editor scoop the first line, clean it up a
little, and make it the title. Since the user has to choose a title
anyway, I thought it might as well be the first line.