Sync preview and editor scroll
This commit makes it so that when the user launches the editor, it
launches at roughly the same scroll position as the preview and puts the
cursor in that area. This is a critical quality-of-life feature because
it saves the user the hassle of managing the scroll positions of both
the editor and the preview independently.

In the future, I hope to add some way to draw the user's eyes to the
position where the cursor lands when the editor launches. At the moment,
it sometimes lands somewhere that's hard for me to see quickly, which I
imagine could get annoying.