Fix two editor bugs
- In Firefox, launching the editor would result in an error complaining
about the non-existence of the editor title field. It turns out that I
was using the value of that field before actually creating it. Weirdly,
this bug didn't seem to bother Chrome or Opera. I think this may have
something to do with the orders in which different browsers load
scripts, but I don't really know. It seems to be fixed now.
- Launching the editor with notes that were too small to have a
scrollbar sometimes resulted in an error. The way I was calculating the
scroll fraction gave NaN for this case, which caused all sorts of
trouble. This value is calculated in a way that accounts for the
no-scroll case now.