Re-add incomplete math support
Math is back. When the user makes changes to the editor that result in
math, the preview tries to render it. `*` and `_` seem to work fine, but
other characters may need to be escaped as well for everything to work
properly.

Note that there is at least one bug, which is that block styling is
still applied in the editor even if the text is between math delimiters.
This is a visual annoyance more than anything, but it seems like it'll
take digging into the parts of CodeMirror that handle Markdown to fix
it, which I'm not up for right now.