Fix a scroll bug with the editor
The new title field in the editor was pushing the rest of the content
down such that CodeMirror wouldn't scroll down until the selection was
several lines below the bottom of the screen. By moving the title field
inside the editor div and making it clickable, this bug is resolved, and
scrolling works as expected.