Add GET /pages/{id} API handler
This commit implements the API handler for getting a single page along with the
user-encrypted key for that page. Other necessary changes include the following:
* Use the Gorilla mux router instead of the standard library one
* Set up a new method of static file handling that is compatible with the new
  router
* Remove the `HttpOnly` flag from the `session_token` cookie (it is important to
  evaluate the site for XSS vulnerabilities)
* Complete the authorization functionality for bearer token authorization

It should be noted that the new router implementation of this commit breaks the
`/page` and `/edit` functionality. Getting this working properly again is not a
priority since a major reason for the API is so that this functionality can be
replaced by asyncronous calls from a single-page app.