Add DELETE /api/pages/{id} functionality
This commit implements the page deletion API handler. This includes the
previously overlooked functionality of deleting all permission entries for the
page to be deleted. A known issue of this branch is that GET /api/pages/{id}
returns HTTP status code 500 for a page which doesn't exist. It should instead
return a 404.